+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -128,7 +128,7 @@ def build_storage_account_resource(name, location, tags, sku):
     return storage_account
 
 
-def build_public_ip_resource(name, location, tags, address_allocation, dns_name=None, zones=None):
+def build_public_ip_resource(name, location, tags, address_allocation, dns_name=None, zone=None):
 
     public_ip_properties = {'publicIPAllocationMethod': address_allocation}
 
@@ -136,7 +136,7 @@ def build_public_ip_resource(name, location, tags, address_allocation, dns_name=
         public_ip_properties['dnsSettings'] = {'domainNameLabel': dns_name}
 
     public_ip = {
-        'apiVersion': '2016-11-01' if zones else '2015-06-15',
+        'apiVersion': '2016-11-01' if zone else '2015-06-15',
         'type': 'Microsoft.Network/publicIPAddresses',
         'name': name,
         'location': location,
@@ -144,8 +144,8 @@ def build_public_ip_resource(name, location, tags, address_allocation, dns_name=
         'dependsOn': [],
         'properties': public_ip_properties
     }
-    if zones:
-        public_ip['zones'] = zones
+    if zone:
+        public_ip['zones'] = zone
 
     return public_ip
 
@@ -307,7 +307,7 @@ def build_vm_resource(  # pylint: disable=too-many-locals
         os_caching=None, data_caching=None, storage_sku=None,
         os_publisher=None, os_offer=None, os_sku=None, os_version=None, os_vhd_uri=None,
         attach_os_disk=None, attach_data_disks=None, data_disk_sizes_gb=None, image_data_disks=None,
-        custom_data=None, secrets=None, license_type=None, zones=None):
+        custom_data=None, secrets=None, license_type=None, zone=None):
 
     def _build_os_profile():
 
@@ -440,8 +440,8 @@ def build_vm_resource(  # pylint: disable=too-many-locals
         'dependsOn': [],
         'properties': vm_properties,
     }
-    if zones:
-        vm['zones'] = zones
+    if zone:
+        vm['zones'] = zone
     return vm