+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -299,7 +299,7 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
                         source=None,  # pylint: disable=unused-argument
                         # below are generated internally from 'source'
                         source_blob_uri=None, source_disk=None, source_snapshot=None,
-                        source_storage_account_id=None, no_wait=False, tags=None, zones=None):
+                        source_storage_account_id=None, no_wait=False, tags=None, zone=None):
     Disk, CreationData, DiskCreateOption = get_sdk(ResourceType.MGMT_COMPUTE, 'Disk', 'CreationData',
                                                    'DiskCreateOption', mod='models')
 
@@ -319,8 +319,8 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
     if size_gb is None and option == DiskCreateOption.empty:
         raise CLIError('usage error: --size-gb required to create an empty disk')
     disk = Disk(location, creation_data, (tags or {}), _get_sku_object(sku), disk_size_gb=size_gb)
-    if zones:
-        disk.zones = [zones]
+    if zone:
+        disk.zones = zone
 
     client = _compute_client_factory()
     return client.disks.create_or_update(resource_group_name, disk_name, disk, raw=no_wait)
@@ -1567,7 +1567,7 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
               storage_account_type=None, vnet_type=None, nsg_type=None, public_ip_type=None, nic_type=None,
               validate=False, custom_data=None, secrets=None, plan_name=None, plan_product=None, plan_publisher=None,
               license_type=None, assign_identity=False, identity_scope=None,
-              identity_role=DefaultStr('Contributor'), identity_role_id=None, zones=None):
+              identity_role=DefaultStr('Contributor'), identity_role_id=None, zone=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, build_vm_resource,
@@ -1585,10 +1585,6 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
     vm_id = resource_id(
         subscription=subscription_id, resource_group=resource_group_name,
         namespace='Microsoft.Compute', type='virtualMachines', name=vm_name)
-
-    # Even for single zonal resources, like vm, services still require an array in payload, so convert here
-    if zones:
-        zones = [zones]
 
     # determine final defaults and calculated values
     tags = tags or {}
@@ -1632,7 +1628,7 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
             master_template.add_resource(build_public_ip_resource(public_ip_address, location,
                                                                   tags,
                                                                   public_ip_address_allocation,
-                                                                  public_ip_address_dns_name, zones))
+                                                                  public_ip_address_dns_name, zone))
 
         subnet_id = subnet if is_valid_resource_id(subnet) else \
             '{}/virtualNetworks/{}/subnets/{}'.format(network_id_template, vnet_name, subnet)
@@ -1685,7 +1681,7 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
         admin_password, ssh_key_value, ssh_dest_key_path, image, os_disk_name,
         os_type, os_caching, data_caching, storage_sku, os_publisher, os_offer, os_sku, os_version,
         os_vhd_uri, attach_os_disk, attach_data_disks, data_disk_sizes_gb, image_data_disks, custom_data, secrets,
-        license_type, zones)
+        license_type, zone)
     vm_resource['dependsOn'] = vm_dependencies
 
     if plan_name: